+++ packages/core/testing/src/testing.ts
@@ -18,6 +18,6 @@ export * from './fake_async';
 export {TestBed, inject, InjectSetupWrapper, withModule} from './test_bed';
 export {TestComponentRenderer, ComponentFixtureAutoDetect, ComponentFixtureNoNgZone, TestModuleMetadata, TestEnvironmentOptions, ModuleTeardownOptions, TestBedStatic} from './test_bed_common';
 export * from './test_hooks';
-export {getTestBed} from './r3_test_bed';
+export {getTestBed} from './test_bed_impl';
 export * from './metadata_override';
 export {MetadataOverrider as ɵMetadataOverrider} from './metadata_overrider';
